+++ BE/girin-grim/src/main/java/com/example/giringrim/university/service/UnivService.java
@@ -1,5 +1,8 @@
-package com.example.giringrim.university;
+package com.example.giringrim.university.service;
 
+import com.example.giringrim.university.repository.UnivRepository;
+import com.example.giringrim.university.entity.University;
+import com.example.giringrim.university.dto.UnivRespDtos;
 import com.example.giringrim.university.exception.RegionNotExistException;
 import com.example.giringrim.university.exception.UnivNotExistException;
 import com.example.giringrim.utils.exception.ErrorMessage;
@@ -26,20 +29,20 @@ public class UnivService {
     }
 
     @Transactional(readOnly = true)
-    public UnivResponseDto.GetUnivListDto getUnivList(String region, String keyword){
+    public UnivRespDtos.GetUnivListDto getUnivList(String region, String keyword){
 
         if(keyword != null){
             List<University> universityList = univRepository.findByKeyword(keyword);
             if(universityList.isEmpty()){
                 throw new UnivNotExistException(ErrorMessage.UNIV_NOT_EXIST);
             }
-            return new UnivResponseDto.GetUnivListDto(universityList);
+            return new UnivRespDtos.GetUnivListDto(universityList);
         }
         List<University> universityList = univRepository.findByRegion(region);
         if(universityList.isEmpty()){
             throw new RegionNotExistException(ErrorMessage.REGION_NOT_EXIST);
         }
-        return new UnivResponseDto.GetUnivListDto(universityList);
+        return new UnivRespDtos.GetUnivListDto(universityList);
     }
